requirements.txt
asgiref==3.8.1
certifi==2024.7.4
Django==5.1
django-elasticsearch-dsl==8.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
